referalapi/requirements.txt
Django==3.2.16
djoser
python-dotenv
drf-spectacular